rag_api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
psycopg2-binary
pgvector
